Refactor and Prep Extensions/Tests
